+++ test.mjs
@@ -1,0 +1,7 @@
+import { LinkedHashMap } from './src/linked-hash-map.mjs';
+
+var lhm = new LinkedHashMap();
+lhm.put("a", "1,2,3");
+console.log(lhm.getAllKeys());
+
+
